+++ config.js
@@ -1,0 +1,3 @@
+module.exports = {
+    PORT: process.env.PORT || 3000
+  };
